requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.3.0
streamlit>=1.38.0
pyyaml>=6.0
python-dotenv>=1.0.0
